Sync ProfileData

+++ BackendServer/resources/InitSQL/1/CreateProfileData.sql
@@ -17,7 +17,7 @@ CREATE TABLE IF NOT EXISTS profile_data (
     current_solo_rank INT NOT NULL,
     highest_team_rank INT NOT NULL,
     division_type TEXT NOT NULL,
-    inventory_version INT NOT NULL DEFAULT 0,
+    inventory_version BIGINT NOT NULL DEFAULT 0,
     crew_id UUID NOT NULL,
     account_id_provider TEXT NOT NULL,
     platform_name TEXT NOT NULL,
